+++ CHANGELOG.md
@@ -48,6 +48,8 @@ release points are not being annotated in GitHub.
 - Sentinel clutter radiometric polynomials are no longer transposed
 - NITFWriter no longer closes passed in file objects
 - Various SICD and SIDD elements fixed to better match NGA standard
+- Fixed case where DTEDInterpolator applied geoid offset incorrectly
+- Improved DTEDInterpolator handling of missing DEMs
 
 ## [1.3.58] - 2023-08-07
 ### Added